requirements.txt
aiogram==2.10.1
aiohttp==3.6.3
aioredis==1.3.1
python-dotenv
